Add JWT validation (#77)
* Add JWT validation

Adds AuthGuard. Includes the request headers in the GraphQL context so the token in the authentication header can be validated.

Signed-off-by: Kurt Johnson <kurt.johnson@headspring.com>

* Create sample env

Signed-off-by: Kurt Johnson <kurt.johnson@headspring.com>

* Remove auth redirects that we won't need

Don't need 'em, delete 'em.

Signed-off-by: Kurt Johnson <kurt.johnson@headspring.com>

* Fetch the Google JWK ourselves

Replace usage of the Google JSON of the JWKS for making a PEM.

Signed-off-by: Kurt Johnson <kurt.johnson@headspring.com>

* Lint auth guard changes

TeamCity threw exceptions because we had some linting issues

Signed-off-by: Kurt Johnson <kurt.johnson@headspring.com>

* Remove unused GOOGLE_JWT_ALGORITHM

Not used. Not needed. Deleted.

Signed-off-by: Kurt Johnson <kurt.johnson@headspring.com>