[FIF-197] Give administrator users ability to upload surveys (#21)
* Add yarn start for ETL

Allow the ETL to start from yarn

* Add migration to add new teacher survey column

Used the same migration as for admin, but renamed for teacher

* Use the new isteachersurveyloader

Update UI, ETL and Api to consume the new column

* Create start tasks for survey and db in ETL

The ETL didn't have start tasks for the database and ETL processors.

* Allow admin and teacher survey loaders

Update navbar component to permit both teachers and admin who can load surveys to see the menu and see the page.

* Add project name version and private to project

Ackages.json was throwing linter warning for SPDX because the packages.json did not have private true. Also added name and version to match other projects

* Allow admin users to load all surveys

Updated SQL for surveyProcessor in ETL to allow teachers and admin each to load surveys

* Update the readme for the new yarn start

README needed an update on ways you can run the etl tasks.