[FIF-225] Convert Student Detail to React (#73)
* Add student detail container

* Save work in progress

Started looking at survey notes and trying to get the bottom purple border to render.

* Add ability to toggle survey and notes areas

* Fix the student contact cards

Scrolling bar width and the max-width

* Load survey info

* Add notes area

* Correct file and component names

Remove Component from names

* Set max width for student detail contact cards

* Fix React key maps

* Update survey bottom border style

* Add notes feature to student detail

* Lint the student detail

* Work in progress on notes

Still not saving notes correctly.

* Lint again

* Fix note editing

* Fix author name

* Add OpenSans font with license

Adding license for open sans

* Use OpenSans on student detail

* Use string for StudentNote.dateadded

We had

* Fix font

Font not font-family

* Update based on PR review

Fixed the phone sizing on surveys. Overrode inherited stylesheets to reclaim border area.

* Update paddings

* Fix issue with add note

* Update delete note

* Show siblings

* Fix button width on student detail

The student detail width was expanding wider than 315px on smaller viewports.

* Correct student card footer style

Fix typo and update the style for the footer to not clash with other footer CSS class names.

* Override h2 bold with font weights

h2-desktop class was too bold on Firefox. Forcing font-weight 400.

* Fix unusual widening of student survey container

The student detail survey container's flexbox was widening in small viewports because we tried to align the chevron with the end of the container, but used 100% widths which expanded the container.

* Fix linting issue

* Hide x overflow in React bodies

We are seeing

* Revert "Fix unusual widening of student survey container"

This reverts commit ed00daace9bb56abd6d00f0cfecfc04c2b1ff69d.

# Conflicts:
#	EdFi.Buzz.UI.Angular/src/globalstyle.ts

* Fix chevron to the right on survey

setting flex-grow 0 moves the chevron to the end.