[FIF-291] Update Deployments for TeamCity (#116)
* Update Deploy for TeamCity

Using Azure Artifacts notes to test on API Deploy

* Add build package execution to BuildAndTestBaseClass

The Kotlin for building and testing needs to also create NuGet package

* Use PR Build template for all Deploy TC builds

Deploy on TeamCity complained that we didn't run PR first, so I switched all dependencies to this.

* Add the database deployment

* Update database project directory casing

The casing has changed for the folder.

* Only deploy to Azure on main branch builds

Branch*Build builds only on default commits - main. This is where we will deploy.