[FIF-347] Add IIS support for UI (#141)
* Restore API install

* Correct the API install path

* Add web.config for React Routes

* Install UI to run as NodeService

Adds the original API Install-NodeService func install.ps1 and updates the XML to start npm. NPM is required on the path now.

* Scope module usage in the current directory

UI doesn't need it. For API,Install-NodeService would run from the database because of scoping the install-module. Made local scope and forced to current directory

* Remove node service install from UI

UI is using IIS. Remove node-service and from testing

* Remove Nginx install funcs

Install-NGinxFiles and install-nginservice not used anywhere

* Remove shared npm variable

We now define it in other files so it was throwing an exception.

* Update readme with IIS install instructions

* Add React friendly IIS web.configs

Routes must be supported

* Update the readme

Removed configuration images after changes, updated UI info.